# Speed up Nano further