--- a/DPGAnalysis/HGCalNanoAOD/test/tb_nano.sh
+++ b/DPGAnalysis/HGCalNanoAOD/test/tb_nano.sh
@@ -7,7 +7,6 @@ fileout=$2
 jobtag=$3
 [[ -z ${jobtag} ]] && jobtag=""
 
-
 cmsDriver.py NANO \
     -s USER:DPGAnalysis/HGCalNanoAOD/hgcRecHits_cff.hgctbTask \
     --datatier NANOAOD \
@@ -21,7 +20,7 @@ cmsDriver.py NANO \
     --era Phase2C17I13M9 \
     --python_filename nanocmsdriver_${jobtag}_cfg.py \
     --customise DPGAnalysis/HGCalTools/tb2023_cfi.configTBConditions \
-    --customise_commands "process.MessageLogger.cerr.FwkReport.reportEvery = 50000 " \
+    --customise_commands "process.NANOAODoutput.compressionAlgorithm = 'ZSTD'\nprocess.NANOAODoutput.compressionLevel = 5\nprocess.MessageLogger.cerr.FwkReport.reportEvery = 50000\nprocess.options.wantSummary = True\n" \
     --no_exec
 
 cmsRun -j FrameworkJobReport_${jobtag}_NANO.xml nanocmsdriver_${jobtag}_cfg.py