~5x speedup with 4 threads:

[Before]
```
TimeReport ---------- Module Summary ---[Real sec]----
TimeReport  per event     per exec    per visit  Name
TimeReport   0.000000     0.000000     0.000000  MEtoEDMConverter
TimeReport   0.001581     0.001581     0.001581  NANOAODoutput
TimeReport   0.000002     0.000002     0.000002  NANOAODoutput_step
TimeReport   0.000001     0.000001     0.000001  TriggerResults
TimeReport   0.000001     0.000001     0.000001  endjob_step
TimeReport   0.000472     0.000472     0.000472  hgcDigiTable
TimeReport   0.000140     0.000140     0.000140  hgctbRecHitsPositionTable
TimeReport   0.000194     0.000194     0.000194  hgctbRecHitsTable
TimeReport   0.000005     0.000005     0.000005  tbMetaDataTable
TimeReport   0.000002     0.000002     0.000002  user_step
TimeReport  per event     per exec    per visit  Name
```

[After]
```
TimeReport ---------- Module Summary ---[Real sec]----
TimeReport  per event     per exec    per visit  Name
TimeReport   0.000000     0.000000     0.000000  MEtoEDMConverter
TimeReport   0.000140     0.000140     0.000140  NANOAODoutput
TimeReport   0.000002     0.000002     0.000002  NANOAODoutput_step
TimeReport   0.000001     0.000001     0.000001  TriggerResults
TimeReport   0.000001     0.000001     0.000001  endjob_step
TimeReport   0.000057     0.000057     0.000057  hgcDigiTable
TimeReport   0.000142     0.000142     0.000142  hgctbRecHitsPositionTable
TimeReport   0.000197     0.000197     0.000197  hgctbRecHitsTable
TimeReport   0.000005     0.000005     0.000005  tbMetaDataTable
TimeReport   0.000002     0.000002     0.000002  user_step
TimeReport  per event     per exec    per visit  Name
```

The output file size is increased by ~2x due to the change from `LZMA:9` to `ZSTD:5`.

@pfs @mariadalfonso